moved encryption to different package
