+++ app/src/main/java/org/fossasia/openevent/app/common/Constants.java
@@ -21,8 +21,6 @@ public final class Constants {
     public static final String PREF_PAYMENT_ACCEPT_ONSITE = "accept_onsite";
     public static final String PREF_PAYMENT_ONSITE_DETAILS = "onsite_details";
     public static final String PREF_PAYMENT_COUNTRY = "key";
-    public static final String PREF_USER_PASSWORD = "user_password";
-    public static final String PREF_USER_EMAIL = "user_email";
 
     private Constants() {
         // Never Called
