+++ app/src/main/java/org/fossasia/openevent/app/common/di/module/ModelModule.java
@@ -8,8 +8,8 @@ import org.fossasia.openevent.app.data.RxBus;
 import org.fossasia.openevent.app.data.auth.AuthService;
 import org.fossasia.openevent.app.data.Bus;
 import org.fossasia.openevent.app.data.SharedPreferencesImpl;
-import org.fossasia.openevent.app.data.auth.EncryptionService;
-import org.fossasia.openevent.app.data.auth.EncryptionServiceImpl;
+import org.fossasia.openevent.app.data.encryption.EncryptionService;
+import org.fossasia.openevent.app.data.encryption.EncryptionServiceImpl;
 import org.fossasia.openevent.app.data.network.ConnectionStatus;
 import org.fossasia.openevent.app.data.network.ConnectionStatusImpl;
